Bug 1906093 - added engineId to avoid duplicated engines - r=gregtatum
Differential Revision: https://phabricator.services.mozilla.com/D215649

+++ toolkit/components/ml/actors/MLEngineChild.sys.mjs
@@ -78,15 +78,39 @@ export class MLEngineChild extends JSWindowActorChild {
           logLevel: lazy.LOG_LEVEL,
         });
 
+        // Check if we already have an engine under this id.
+        if (this.#engineDispatchers.has(options.engineId)) {
+          lazy.console.debug("EngineDispatcher already exists");
+          let currentEngineDispatcher = this.#engineDispatchers.get(
+            options.engineId
+          );
+
+          // The option matches, let's reuse the engine
+          if (currentEngineDispatcher.pipelineOptions.equals(options)) {
+            lazy.console.debug(
+              "PipelineOptions are equal, reusing the same engine"
+            );
+            return;
+          }
+
+          // The options do not match, terminate the old one so we have a single engine per id.
+          lazy.console.debug(
+            "PipelineOptions are not equal, terminating the old one"
+          );
+
+          await currentEngineDispatcher.terminate(/* shutDownIfEmpty */ false);
+          this.#engineDispatchers.delete(options.engineId);
+        }
+
         this.#engineDispatchers.set(
-          options.taskName,
+          options.engineId,
           new EngineDispatcher(this, port, options)
         );
         break;
       }
       case "MLEngine:ForceShutdown": {
         for (const engineDispatcher of this.#engineDispatchers.values()) {
-          return engineDispatcher.terminate();
+          await engineDispatcher.terminate(/* shutDownIfEmpty */ true);
         }
         this.#engineDispatchers = null;
         break;
@@ -117,16 +141,29 @@ export class MLEngineChild extends JSWindowActorChild {
    * @returns {Promise<object>}
    */
   getInferenceOptions(taskName) {
-    return this.sendQuery(`MLEngine:GetInferenceOptions:${taskName}`);
+    return this.sendQuery("MLEngine:GetInferenceOptions", {
+      taskName,
+    });
   }
 
   /**
-   * @param {string} engineName
+   * Removes an engine by its ID. Optionally shuts down if no engines remain.
+   *
+   * @param {string} engineId - The ID of the engine to remove.
+   * @param {boolean} [shutDownIfEmpty] - If true, shuts down the engine process if no engines remain.
    */
-  removeEngine(engineName) {
-    this.#engineDispatchers.delete(engineName);
-    if (this.#engineDispatchers.size === 0) {
-      this.sendQuery("MLEngine:DestroyEngineProcess");
+  removeEngine(engineId, shutDownIfEmpty) {
+    if (!this.#engineDispatchers) {
+      return;
+    }
+    lazy.console.debug(`Removing engine ${engineId}`);
+    this.#engineDispatchers.delete(engineId);
+    this.sendAsyncMessage("MLEngine:Removed", { engineId });
+    if (this.#engineDispatchers.size === 0 && shutDownIfEmpty) {
+      lazy.console.debug(
+        "Shutting down by calling MLEngine:DestroyEngineProcess"
+      );
+      this.sendAsyncMessage("MLEngine:DestroyEngineProcess");
     }
   }
 }
@@ -151,7 +188,14 @@ class EngineDispatcher {
   /** @type {string} */
   #taskName;
 
-  /** Creates the inference engine given the wasm runtime and the run options.
+  /** @type {string} */
+  #engineId;
+
+  /** @type {PipelineOptions | null} */
+  pipelineOptions = null;
+
+  /**
+   * Creates the inference engine given the wasm runtime and the run options.
    *
    * The initialization is done in three steps:
    * 1. The wasm runtime is fetched from RS
@@ -175,6 +219,8 @@ class EngineDispatcher {
     mergedOptions.updateOptions(pipelineOptions);
     lazy.console.debug("Inference engine options:", mergedOptions);
 
+    this.pipelineOptions = mergedOptions;
+
     return InferenceEngine.create({
       wasm,
       pipelineOptions: mergedOptions,
@@ -191,6 +237,7 @@ class EngineDispatcher {
     this.mlEngineChild = mlEngineChild;
     this.#taskName = pipelineOptions.taskName;
     this.timeoutMS = pipelineOptions.timeoutMS;
+    this.#engineId = pipelineOptions.engineId;
 
     this.#engine = this.initializeInferenceEngine(
       pipelineOptions,
@@ -264,7 +311,7 @@ class EngineDispatcher {
           break;
         }
         case "EnginePort:Terminate": {
-          this.terminate();
+          await this.terminate(/* shutDownIfEmpty */ data.shutdown);
           break;
         }
         case "EnginePort:ModelResponse": {
@@ -296,7 +343,7 @@ class EngineDispatcher {
               error,
             });
             // The engine failed to load. Terminate the entire dispatcher.
-            this.terminate();
+            await this.terminate(/* shutDownIfEmpty */ true);
             return;
           }
 
@@ -330,23 +377,30 @@ class EngineDispatcher {
 
   /**
    * Terminates the engine and its worker after a timeout.
+   *
+   * @param {boolean} shutDownIfEmpty - If true, shuts down the engine process if no engines remain.
    */
-  async terminate() {
+  async terminate(shutDownIfEmpty) {
     if (this.#keepAliveTimeout) {
       lazy.clearTimeout(this.#keepAliveTimeout);
       this.#keepAliveTimeout = null;
     }
     if (this.#port) {
+      lazy.console.log(
+        "Terminating the engine by calling EnginePort:EngineTerminated"
+      );
       // This call will trigger back an EnginePort:Discard that will close the port
       this.#port.postMessage({ type: "EnginePort:EngineTerminated" });
+    } else {
+      lazy.console.log("No more port to terminate the engine");
     }
-    this.mlEngineChild.removeEngine(this.#taskName);
     try {
       const engine = await this.#engine;
       engine.terminate();
     } catch (error) {
       lazy.console.error("Failed to get the engine", error);
     }
+    this.mlEngineChild.removeEngine(this.#engineId, shutDownIfEmpty);
   }
 }
 
@@ -452,7 +506,9 @@ class InferenceEngine {
   }
 
   terminate() {
-    this.#worker.terminate();
-    this.#worker = null;
+    if (this.#worker) {
+      this.#worker.terminate();
+      this.#worker = null;
+    }
   }
 }
